draws geospatial map on matplotlib with precise mapping of colors, like in the other project written in R.
(There was a small visual issue because a same color, let's say "blue",
wasn't understood with the same tonality of the color between the
geospatial map provider -Google Maps- and the legend of colors and
transit direction written in Python matplotlib. E.g., following for the
case of "blue", if one understood as RGB 0000FF, the other understood
it as RGB 0000FE, which is a tiny change in the tonality. This affected
the legend of the directions of the transit vehicles. Now both use
explicit RGB hexadecimal values instead of color names. E.g., explicitly
0000FF instead of 'blue', for both understood different things for this
term.)