moving geospatial check from matplotib up to xml parsing of data frame
Before, the geospatial check was done in the visualization of the Google
Map inside matplotlib, and this could cause issues with the color legend
when a color was pre-assigned in the legend but later left unused as there
happened to be no transit bus inside that geospatial area (ie., the
color was pre-allocated in the legend before checking if it was really to
be used). Now, the check of whether the transit vehicle is inside the
geospatial area is done directly in the parsing of the NextBus real-time
location XML into a Panda Data frame, ie., the Panda data frame only has
those interesting vehicles in this geospatial area.

Another benefit of this is that the center of the Google Maps is exactly
the center of the needed vehicles, not the center of all vehicles without
having applied the geospatial condition, for the later geospatial center
may be shifted in the space, and hence the resulting Google Maps appearing
shifted or a little out-of-focus. This is corrected now.